Rename Crc32 to Crc32c
