+++ src-c/native/jbuild
@@ -1,6 +1,6 @@
 (library
  ((name laolao)
   (public_name checkseum.laolao)
-  (c_names (stubs adler32 crc32))
+  (c_names (stubs adler32 crc32c))
   (c_library_flags (-lpthread))
   (c_flags (-I.))))
